requirements.txt
Django==2.0
django-jet==1.0.7
django-rosetta==0.8.1
mysqlclient==1.3.10
pycodestyle==2.3.1
simplejson==3.13.2
